pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sssn"
version = "0.1.0"
description = "Simple System of Systems Network"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Productive Superintelligence" },
]
dependencies = [
    "fastapi>=0.110",
    "pydantic>=2.0",
    "PyJWT>=2.8",
    "uvicorn>=0.29",
    "aiosqlite>=0.20",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.urls]
Homepage = "https://github.com/Productive-Superintelligence/sssn"
Repository = "https://github.com/Productive-Superintelligence/sssn"

[tool.setuptools.packages.find]
include = ["sssn*"]
